+++ test/resources/any-type-doc-definitions.js
@@ -1,0 +1,27 @@
+function() {
+  return {
+    anyTypeDoc: {
+      typeFilter: simpleTypeFilter,
+      channels: { write: 'write' },
+      propertyValidators: {
+        arrayProp: {
+          type: 'array',
+          arrayElementsValidator: {
+            type: 'any',
+            required: true
+          }
+        },
+        hashtableProp: {
+          type: 'hashtable',
+          hashtableValuesValidator: {
+            type: 'any',
+            immutableWhenSet: true
+          }
+        },
+        anyProp: {
+          type: 'any'
+        }
+      }
+    }
+  };
+}